// src/app/shared/types/api-types/bidt-product-customer-dto.interface.ts
export interface BidtProductCustomerDTO {
  bidoCustomerCustomerCode?: string;
  productQuantity?: string;
  numberOfDays?: number;
  priceValue?: string;
  priceUnit?: string;
  endOfValidityDate?: number;
  id?: number;
  supplierName?: string;
  serviceType?: string;
  bidoCustomerMandant?: string;
  bidoCustomerName?: string;
}